+++ src/main/java/gift/model/wishlist/WishResponse.java
@@ -1,2 +1,11 @@
-package gift.model.wishlist;public record WishResponse() {
+package gift.model.wishlist;
+
+import gift.model.product.Product;
+
+public record WishResponse(Long id, Long productId, String Name, int count) {
+    public static WishResponse from(WishList wishList, Product product) {
+        return new WishResponse(wishList.getId(), product.getId(), product.getName(),
+            wishList.getQuantity());
+    }
+
 }
